requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==3.2.25
django-cache-url==3.4.5
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-environ==0.11.2
django-js-asset==2.2.0
djangorestframework==3.13.1
environs==8.0.0
marshmallow==3.21.3
mysqlclient==2.2.4
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
